requirements.txt
kivy>=2.3.0
python-chess>=1.10.0
